# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CampusPulse API
  version: 1.0.0
  description: Production API for CampusPulse student platform
servers:
  - url: /api
paths:
...
  /products/{id}:
    get:
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Product details }
...